ah_rd_domains.h: quiet shift warnings
The W0 and W1 macros use bounds checks to ensure that their shifts
are safe in the sense of the shift count being non-negative and not
greater than the type width.

Unfortuantely clang still warns about code like this
	int x = 1 ? 0 : (1 << -1);
	int y = 1 ? 0 : (1 << 64);
even though the negative or overflowing shift won't actually be
generated.  Consequently clang generates warnings for hundreds of uses
of W0 and W1, even though they are actually safe.

That clang warns about this is a bug that has been known for at least
six years [1] and these specific warnings have been discussed on the
lists before [2].  Perhaps it is time we stopped waiting on clang and
just worked around the bogus warnings, so that we don't lose sight of
useful warnings.

This patch makes it so that the compiler doesn't actually see the
negative or overflowing shift by repeating the same bounds check on the
RHS of the shift and hiding the would-be bad values.

[1] https://bugs.llvm.org/show_bug.cgi?id=10030
[2] https://lists.freebsd.org/pipermail/freebsd-toolchain/2011-November/000277.html

Reviewers: adrian, dim
Subscribers: arundel, markj