malloc/zalloc M_NOWAIT failure injection
The MALLOC_MAKE_FAILURES kernel option could be used to inject failures
for malloc(9) allocations which use the M_NOWAIT flag.  This is an
expansion and enhancement.
 - Failures may now be injected for UMA zalloc instead of just malloc.
 - A fail(9) fail_point now controls injection instead of an ad hoc rate
   mechanism.
 - A whitelist and blacklist now allow specific malloc type and UMA zone
   names to be targeted or avoided.
 - Details about the last injection are now recorded to aid debugging.

This currently still lacks a manual page and the option is not yet
enabled in any kernel configurations.

Reviewers: markj, rwatson
Subscribers: glebius, jeff

Sponsored by:	Dell EMC Isilon
Differential Revision:	https://reviews.freebsd.org/D20714