igmp: solve M_NOWAIT softc allocation failure
The igmp_ifsoftc allocation was done with M_NOWAIT, but it did not
handle failure.  It may have trapped immediately, or at detach time.
Actually, the allocation does not appear to need to be M_NOWAIT at all,
except for that it was being done under a mutex.  Instead, do an
M_WAITOK allocation outside the mutex.

Sponsored by:	Dell EMC Isilon

Reviewers: glebius
Subscribers: mmacy