rtld-elf: link static no-SSP libcompiler_rt
Link rtld-elf with a special static build of libcompiler_rt without
-fstack-protector (SSP).  Compilers may emit calls to builtin functions,
but because of rtld's special needs and the use of the linker flag
-nostdlib, the compiler builtin functions are not linked as normal.
This now links the builtin library but built in a way compatible with
rtld-elf.

Previously we had two different solutions.  On some architectures, we
were able to extract reimplementations of the necessary builtin
functions from our special build of libc.  On ARM, we linked
libcompiler_rt, but didn't guarantee that it didn't use SSP.

This is motivated by the same issue as D26199 and D27665, but should be
a simpler solution that will apply to all architectures.

Differential Revision:	https://reviews.freebsd.org/D27736