rtld-elf: just link libcompiler_rt
Compilers may emit calls to builtin functions, but because of rtld's
special needs and the use of the linker flag -nostdlib, the compiler
builtin functions are not linked as normal.  This now links the builtin
library.

Technically it is not correct to link rtld with the normal
libcompiler_rt, as rtld is not compatible with -fstack-protector (SSP)
but libcompiler_rt is compiled with it.  However, in practice, we
haven't yet seen SSP be used on any of the objects rtld needs from
libcompiler_rt (which so far is just some extended width math
functions).  Rather than increasing the build time by rebuilding the
library, just link the normal one for now and plan to revist this if it
ever actually causes a problem.

Previously we had two different solutions.  On some architectures, we
were able to extract reimplementations of the necessary builtin
functions from our special build of libc.  On ARM, we just linked
libcompiler_rt.

This is motivated by the same issue as D26199 and D27665, but should be
a simpler solution that will apply to all architectures.

Differential Revision:	https://reviews.freebsd.org/D27736