lockmgr: introduce LK_CONDWAIT
The idea is that a locker can safely sleep without an interlock if it
registers as a holder by incrementing an atomic counter associated with
the lock.  After taking the sleepq lock but before going to sleep, we
check that the sleep condition is still current via a callback from
lockmgr.  If not, we skip the sleep and return ENOLCK.