lockmgr: introduce LK_CONDWAIT
The idea is that a locker can safely wait without an interlock by
checking the wait condition via a callback from lockmgr under the
sleepq lock, after the lock's waiter flag is set.  If not, we skip the
sleep and return ENOLCK.  A cooperating wait invalidator calls checks
the waiter bits and does a wakeup after invalidating the wait
conditions.