rb tree: quiet gcc -Wmaybe-uninitialized
The comment explains that the variable should never be used
uninitialized according to the algorithm, but the compiler can't be
expected to infer that.

There are many GCC -Wmaybe-unintialized warnings in the build, but this
one was particularly noisy due to being in a system header and due to
being triggered via multiple function inlining paths.  This squelches
around 250 warning messages.

Differential Revision:	https://reviews.freebsd.org/D52940